+++ tests/test_data_store_populate.py
@@ -46,14 +46,14 @@ class DataStorePopulateSpatiaLiteTestCase(TestCase):
         with self.store.session_scope():
             nationalities = self.store.session.query(self.store.db_classes.Nationality).all()
             platform_types = self.store.session.query(self.store.db_classes.PlatformType).all()
-            nationality_object = self.store.search_nationality("UK")
+            nationality_object = self.store.search_nationality("United Kingdom")
             platform_type_object = self.store.search_platform_type("PLATFORM-TYPE-1")
 
             # Check whether they are not empty anymore and filled with correct data
             self.assertNotEqual(len(nationalities), 0)
             self.assertNotEqual(len(platform_types), 0)
 
-            self.assertIn(nationality_object.name, "UK")
+            self.assertIn(nationality_object.name, "United Kingdom")
             self.assertIn(platform_type_object.name, "PLATFORM-TYPE-1")
 
     def test_populate_metadata(self):
@@ -99,7 +99,7 @@ class DataStorePopulateSpatiaLiteTestCase(TestCase):
                 .filter_by(nationality_id=platform_object.nationality_id)
                 .first()
             )
-            self.assertEqual(nationality.name, "UK")
+            self.assertEqual(nationality.name, "United Kingdom")
             platform_type = (
                 self.store.session.query(self.store.db_classes.PlatformType)
                 .filter_by(platform_type_id=platform_object.platform_type_id)
@@ -188,14 +188,14 @@ class DataStorePopulatePostGISTestCase(TestCase):
         with self.store.session_scope():
             nationalities = self.store.session.query(self.store.db_classes.Nationality).all()
             platform_types = self.store.session.query(self.store.db_classes.PlatformType).all()
-            nationality_object = self.store.search_nationality("UK")
+            nationality_object = self.store.search_nationality("United Kingdom")
             platform_type_object = self.store.search_platform_type("PLATFORM-TYPE-1")
 
             # Check whether they are not empty anymore and filled with correct data
             self.assertNotEqual(len(nationalities), 0)
             self.assertNotEqual(len(platform_types), 0)
 
-            self.assertIn(nationality_object.name, "UK")
+            self.assertIn(nationality_object.name, "United Kingdom")
             self.assertIn(platform_type_object.name, "PLATFORM-TYPE-1")
 
     def test_populate_metadata(self):
@@ -241,7 +241,7 @@ class DataStorePopulatePostGISTestCase(TestCase):
                 .filter_by(nationality_id=platform_object.nationality_id)
                 .first()
             )
-            self.assertEqual(nationality.name, "UK")
+            self.assertEqual(nationality.name, "United Kingdom")
             platform_type = (
                 self.store.session.query(self.store.db_classes.PlatformType)
                 .filter_by(platform_type_id=platform_object.platform_type_id)