+++ HISTORY.rst
@@ -2,6 +2,19 @@
 History
 =======
 
+0.0.13 (2020-05-28)
+------------------
+
+* Offer default vaues in Command Line Resolver
+* Export snapshot of database to SQLite
+
+0.0.11 (2020-05-20)
+------------------
+
+* Use caching to improve import performance
+* Use database migration to allow updates to "live" databases
+* Add ability to view database contents from Admin interface
+
 0.0.9 (2020-04-04)
 ------------------